[GHI #36] Improved Importing and documentation (#55)
This PR aims to simplify the importing process while allowing flexibility and clarity in the documentation. It removes tom select, modal, and panel from being imported by default. It adds the ability to import tokens or components as folders instead of specifying each file. It defines a structure for how add-ons can be created and integrated into various applications. It also adds clearer documentation around Selective Importing and new documentation on the Addons.