[TR#22] Flash Alert Revisions (#158)
This PR revises the Flash and Alert components by combining them into one Alert. It also changes the alert to be outlined by default and filled as a modifier. It removes the old flash component, and adds a flash option to the alert component. It also fixed the documentation code view formatting.